backend-example/requirements.txt
flask==2.3.0
flask-cors==4.0.0
pandas==2.0.0
scikit-learn==1.3.0
joblib==1.3.0
pydantic==2.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
